Submit working game in web page
A basic web page and style have been created
(index.html and style.css) and the JS code has
been modified to work with the page. The game
is now functional on the page. The next step is
to complete the styling of the page to make it
more presentable.